+++ contracts/riskmgmt/RiskMgmt.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.4.11;
 
-import './RiskMgmtAdapter.sol';
+import './RiskMgmtInterface.sol';
 
 /// @title RiskMgmt Contract
 /// @author Melonport AG <team@melonport.com>
-contract RiskMgmt is RiskMgmtAdapter {
+contract RiskMgmt is RiskMgmtInterface {
 
     // FIELDS
 
@@ -17,10 +17,10 @@ contract RiskMgmt is RiskMgmtAdapter {
     // NON-CONSTANT METHODS
 
     function isExchangeMakePermitted(
-        ERC20    haveToken,
-        ERC20    wantToken,
-        uint128  haveAmount,
-        uint128  wantAmount
+        ERC20   haveToken,
+        ERC20   wantToken,
+        uint    haveAmount,
+        uint    wantAmount
     )
         returns (bool)
     {
@@ -28,10 +28,10 @@ contract RiskMgmt is RiskMgmtAdapter {
     }
 
     function isExchangeTakePermitted(
-        ERC20    haveToken,
-        ERC20    wantToken,
-        uint128  haveAmount,
-        uint128  wantAmount,
+        ERC20   haveToken,
+        ERC20   wantToken,
+        uint    haveAmount,
+        uint    wantAmount,
         address orderOwner
     )
         returns (bool)