+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Vault internal tracking of all orders
 - Function `subscribeRequest` in Vault
 - `AssetRegistrar` Contract
+- Function `getDataHistory` in PriceFeed
 
 ### Changed
 - Use Logging contract from Vault